Add env var to skip deleting stemcells in integration specs
- Common iteration pattern with integration specs is to upload a
  stemcell _once_, then use it subsequently by setting the
  BOSH_VSPHERE_STEMCELL_ID. During that initial run, one must comment
  out the `delete_stemcell` directive in the `after` cleanup block. If
  one were to check this in accidentially it would create a mess in our
  testbeds. Adding this ENV var allows one to avoid changing the code
  and therefore reduces the possiblity of an accidential check-in.

+++ src/vsphere_cpi/spec/integration/spec_helper.rb
@@ -47,7 +47,7 @@ RSpec.configure do |rspec_config|
     setup_global_config
     fetch_global_properties
 
-    if ENV.fetch('BOSH_VSPHERE_STEMCELL_ID', '').empty?
+    if ENV.fetch('BOSH_VSPHERE_STEMCELL_ID', '').empty? && ENV.fetch('SKIP_STEMCELL_DELETION', '').empty?
       delete_stemcell(@lifecycle_cpi, stemcell_id)
     end
     @lifecycle_cpi.cleanup
